test(slr_jest) pull HTML from PHP snapshots
Implement a first version of a Jest utility creating HTML documents for
the `slr_jest` test suite pulling the HTML snapshots generated from the
`slr_integration` suite.
